// src/models/ToDo.ts
export interface ToDo {
    assigned_by: string;
    creation: string;
    doctype: string;
    owner: string;
    modified_by: string;
    reference_type: string;
    priority: string;
    docstatus: number;
    status: string;
    description: string;
    reference_name: string;
    assigned_by_full_name: string;
    date: string;
    name: string;
    idx: number;
    modified: string;
}